+++ modules/Custom/MyReport.php
@@ -168,12 +168,14 @@ if ( ! $_REQUEST['modfunc'] )
 
 		if ( empty( $_REQUEST['LO_save'] ) )
 		{
-			$_SESSION['List_PHP_SELF'] = PreparePHP_SELF( $_REQUEST, [ 'bottom_back' ] );
+			/**
+			 * Remove need to make an AJAX call to Bottom.php
+			 *
+			 * @since 12.0 JS Show BottomButtonBack & update its URL & text
+			 */
+			require_once 'ProgramFunctions/Bottom.fnc.php';
 
-			$_SESSION['Back_PHP_SELF'] = 'student';
-
-			// Update Bottom.php.
-			echo '<script>ajaxLink("Bottom.php"); old_modname="";</script>';
+			BottomButtonBackUpdate( 'student' );
 		}
 
 		ListOutput( $students_RET, $LO_columns, 'Student', 'Students', false, $LO_group );