+++ modules/Students/Search.inc.php
@@ -313,15 +313,17 @@ else
 
 		if ( empty( $_REQUEST['LO_save'] ) && empty( $extra['suppress_save'] ) )
 		{
-			$_SESSION['List_PHP_SELF'] = PreparePHP_SELF( $_REQUEST, [ 'bottom_back' ] );
-
-			$_SESSION['Back_PHP_SELF'] = 'student';
-
 			if ( User( 'PROFILE' ) === 'admin'
 				|| User( 'PROFILE' ) === 'teacher' )
 			{
-				// Update Bottom.php.
-				echo '<script>ajaxLink("Bottom.php"); old_modname="";</script>';
+				/**
+				 * Remove need to make an AJAX call to Bottom.php
+				 *
+				 * @since 12.0 JS Show BottomButtonBack & update its URL & text
+				 */
+				require_once 'ProgramFunctions/Bottom.fnc.php';
+
+				BottomButtonBackUpdate( 'student' );
 			}
 		}