+++ modules/misc/Export.php
@@ -521,21 +521,17 @@ if ( isset( $_REQUEST['search_modfunc'] )
 
 		$RET = GetStuList( $extra );
 
-		if ( ! empty( $extra['array_function'] )
-			&& function_exists( $extra['array_function'] ) )
-		{
-			$extra['array_function']( $RET );
-		}
-
 		if ( empty( $_REQUEST['LO_save'] )
 			&& empty( $extra['suppress_save'] ) )
 		{
-			$_SESSION['List_PHP_SELF'] = PreparePHP_SELF( $_REQUEST, [ 'bottom_back' ] );
+			/**
+			 * Remove need to make an AJAX call to Bottom.php
+			 *
+			 * @since 12.0 JS Show BottomButtonBack & update its URL & text
+			 */
+			require_once 'ProgramFunctions/Bottom.fnc.php';
 
-			$_SESSION['Back_PHP_SELF'] = 'student';
-
-			// Update Bottom.php.
-			echo '<script>ajaxLink("Bottom.php"); old_modname="";</script>';
+			BottomButtonBackUpdate( 'student' );
 		}
 
 		if ( ! isset( $_REQUEST['_ROSARIO_PDF'] ) )